Remove unneccessary files and organize project file
