+++ OmiseSwift/BankAccount.swift
@@ -1,18 +1,53 @@
 import Foundation
 
-public class BankAccount: ResourceObject {
-    public var brand: String? {
-        get { return get("brand", StringConverter.self) }
-        set { set("brand", StringConverter.self, toValue: newValue) }
-    }
+public struct BankAccount: OmiseObject {
+    public let object: String
     
-    public var lastDigits: LastDigits? {
-        get { return get("last_digits", LastDigitsConverter.self) }
-        set { set("last_digits", LastDigitsConverter.self, toValue: newValue) }
-    }
+    public let brand: String
+    public let accountNumber: String
+    public let lastDigits: LastDigits
+    public let name: String
+}
+
+extension BankAccount {
+    public init?(JSON json: Any) {
+        guard let json = json as? [String: Any],
+            let object = BankAccount.parseObject(JSON: json),
+            let brand = json["brand"] as? String,
+            let name = json["name"] as? String,
+            let accountNumber = json["number"] as? String,
+            let lastDigits = json["last_digits"].flatMap(LastDigitsConverter.convert(fromAttribute:)) else {
+                return nil
+        }
         
-    public var name: String? {
-        get { return get("name", StringConverter.self) }
-        set { set("name", StringConverter.self, toValue: newValue) }
+        self.object = object
+        self.brand = brand
+        self.name = name
+        self.accountNumber = accountNumber
+        self.lastDigits = lastDigits
     }
 }
+
+public struct BankAccountParams: APIParams {
+    public let brand: String?
+    public let accountNumber: String?
+    public let name: String?
+
+    public var json: JSONAttributes {
+        return Dictionary<String, Any>.makeFlattenDictionaryFrom([
+            "brand": brand,
+            "number": accountNumber,
+            "name": name,
+        ])
+    }
+}
+
+
+extension BankAccountParams {
+    public init(createNewBankAccountWithBrand brand: String, accountNumber: String, name: String) {
+        self.brand = brand
+        self.name = name
+        self.accountNumber = accountNumber
+    }
+}
+
