+++ OmiseSwift/Card.swift
@@ -1,134 +1,126 @@
 import Foundation
 
-open class Card: ResourceObject {
-    open override class var info: ResourceInfo {
-        return ResourceInfo(parentType: Customer.self, path: "/cards")
-    }
+public enum CardBrand: String {
+    case visa = "Visa"
+    case masterCard = "MasterCard"
+    case jcb = "JCB"
+    case amex = "American Express"
+    case diners = "Diners Club"
+    case discover = "Discover"
+    case maestro = "Maestro"
+}
+
+
+public enum CardFinancing: String {
+    case credit
+    case debit
+}
+
+
+public struct Card: OmiseResourceObject {
+    public static let resourceInfo: ResourceInfo = ResourceInfo(parentType: Customer.self, path: "/cards")
     
-    public var country: String? {
-        get { return get("country", StringConverter.self) }
-        set { set("country", StringConverter.self, toValue: newValue) }
-    }
+    public let location: String
+    public let object: String
     
-    public var city: String? {
-        get { return get("city", StringConverter.self) }
-        set { set("city", StringConverter.self, toValue: newValue) }
-    }
+    public let id: String
+    public let isLive: Bool
+    public var createdDate: Date
+    public let isDeleted: Bool
     
-    public var bank: String? {
-        get { return get("bank", StringConverter.self) }
-        set { set("bank", StringConverter.self, toValue: newValue) }
-    }
+    public let countryCode: String?
+    public let city: String?
+    public let postalCode: String?
     
-    public var postalCode: String? {
-        get { return get("postal_code", StringConverter.self) }
-        set { set("postal_code", StringConverter.self, toValue: newValue) }
-    }
+    public let bankName: String?
     
-    public var financing: String? {
-        get { return get("financing", StringConverter.self) }
-        set { set("financing", StringConverter.self, toValue: newValue) }
-    }
+    public let lastDigits: LastDigits
+    public let brand: CardBrand
+    public let expiration: (month: Int, year: Int)?
     
-    public var lastDigits: LastDigits? {
-        get { return get("last_digits", LastDigitsConverter.self) }
-        set { set("last_digits", LastDigitsConverter.self, toValue: newValue) }
-    }
+    public let name: String?
+    public let fingerPrint: String
     
-    public var brand: CardBrand? {
-        get { return get("brand", EnumConverter.self) }
-        set { set("brand", EnumConverter.self, toValue: newValue) }
-    }
-    
-    public var expirationMonth: Int? {
-        get { return get("expiration_month", IntConverter.self) }
-        set { set("expiration_month", IntConverter.self, toValue: newValue) }
-    }
-    
-    public var expirationYear: Int? {
-        get { return get("expiration_year", IntConverter.self) }
-        set { set("expiration_year", IntConverter.self, toValue: newValue) }
-    }
-    
-    public var fingerprint: String? {
-        get { return get("fingerprint", StringConverter.self) }
-        set { set("fingerprint", StringConverter.self, toValue: newValue) }
-    }
-    
-    public var name: String? {
-        get { return get("name", StringConverter.self) }
-        set { set("name", StringConverter.self, toValue: newValue) }
-    }
-    
-    public var securityCodeCheck: Bool? {
-        get { return get("security_code_check", BoolConverter.self) }
-        set { set("security_code_check", BoolConverter.self, toValue: newValue) }
+    public let financing: CardFinancing?
+}
+
+
+extension Card {
+    public init?(JSON json: Any) {
+        guard let json = json as? [String: Any] else {
+            return nil
+        }
+        
+        guard let omiseObjectProperties = Card.parseOmiseResource(JSON: json),
+            let lastDigits = LastDigitsConverter.convert(fromAttribute: json["last_digits"]),
+            let cardBrand: CardBrand = EnumConverter.convert(fromAttribute: json["brand"]),
+            let fingerPrint = json["fingerprint"] as? String else {
+                return nil
+        }
+        
+        (self.object, self.location, self.id, self.isLive, self.createdDate, self.isDeleted) = omiseObjectProperties
+        self.lastDigits = lastDigits
+        self.brand = cardBrand
+        
+        self.name = json["name"] as? String
+        self.bankName = json["bank"] as? String
+        self.postalCode = json["postal_code"] as? String
+        self.countryCode = json["country"] as? String
+        self.city = json["city"] as? String
+        
+        self.financing = EnumConverter.convert(fromAttribute: json["financing"])
+        self.fingerPrint = fingerPrint
+        
+        if let expirationMonth = json["expiration_month"] as? Int, let expirationYear = json["expiration_year"] as? Int {
+            self.expiration = (month: expirationMonth, year: expirationYear)
+        } else {
+            self.expiration = nil
+        }
     }
 }
 
-public class CardParams: Params {
-    public var name: String? {
-        get { return get("name", StringConverter.self) }
-        set { set("name", StringConverter.self, toValue: newValue) }
-    }
-    
-    public var expirationMonth: Int? {
-        get { return get("expiration_month", IntConverter.self) }
-        set { set("expiration_month", IntConverter.self, toValue: newValue) }
-    }
-    
-    public var expirationYear: Int? {
-        get { return get("expiration_year", IntConverter.self) }
-        set { set("expiration_year", IntConverter.self, toValue: newValue) }
-    }
-    
-    public var postalCode: String? {
-        get { return get("postal_code", StringConverter.self) }
-        set { set("postal_code", StringConverter.self, toValue: newValue) }
-    }
-    
-    public var city: String? {
-        get { return get("city", StringConverter.self) }
-        set { set("city", StringConverter.self, toValue: newValue) }
+public struct CardParams: APIParams {
+    public var name: String?
+    public var expirationMonth: Int?
+    public var expirationYear: Int?
+    public var postalCode: String?
+    public var city: String?
+
+    public var json: JSONAttributes {
+        return Dictionary.makeFlattenDictionaryFrom([
+            "name": name,
+            "expiration_month": expirationMonth,
+            "expiration_year": expirationYear,
+            "postal_code": postalCode,
+            "city": city,
+        ])
     }
 }
+
 
 extension Card: Listable { }
 extension Card: Retrievable { }
 extension Card: Destroyable { }
-
 extension Card: Updatable {
     public typealias UpdateParams = CardParams
 }
 
+
 extension Customer {
-    public func listCards(using given: Client? = nil, params: ListParams? = nil, callback: @escaping Card.ListOperation.Callback) -> Request<Card.ListOperation.Result>? {
-        return Card.list(using: given ?? attachedClient, parent: self, params: params, callback: callback)
+    public func listCards(using client: APIClient, params: ListParams? = nil, callback: @escaping Card.ListRequest.Callback) -> Card.ListRequest? {
+        return Card.list(using: client, parent: self, params: params, callback: callback)
     }
     
-    public func retrieveCard(using given: Client? = nil, id: String, callback: @escaping Card.RetrieveOperation.Callback) -> Request<Card.RetrieveOperation.Result>? {
-        return Card.retrieve(using: given ?? attachedClient, parent: self, id: id, callback: callback)
+    public func retrieveCard(using client: APIClient, id: String, callback: @escaping Card.RetrieveRequest.Callback) -> Card.RetrieveRequest? {
+        return Card.retrieve(using: client, parent: self, id: id, callback: callback)
     }
     
-    public func updateCard(using given: Client? = nil, id: String, params: CardParams, callback: @escaping Card.UpdateOperation.Callback) -> Request<Card.UpdateOperation.Result>? {
-        return Card.update(using: given ?? attachedClient, parent: self, id: id, params: params, callback: callback)
+    public func updateCard(using client: APIClient, id: String, params: CardParams, callback: @escaping Card.UpdateRequest.Callback) -> Card.UpdateRequest? {
+        return Card.update(using: client, parent: self, id: id, params: params, callback: callback)
     }
     
-    public func destroyCard(using given: Client? = nil, id: String, callback: @escaping Card.DestroyOperation.Callback) -> Request<Card.DestroyOperation.Result>? {
-        return Card.destroy(using: given ?? attachedClient, parent: self, id: id, callback: callback)
+    public func destroyCard(using client: APIClient, id: String, callback: @escaping Card.DestroyRequest.Callback) -> Card.DestroyRequest? {
+        return Card.destroy(using: client, parent: self, id: id, callback: callback)
     }
 }
 
-func exampleCard() {
-    let customer = Customer()
-    customer.id = "cust_test_123"
-    
-    _ = Card.list(parent: customer) { (result) in
-        switch result {
-        case let .success(cards):
-            print("cards: \(cards)")
-        case let .fail(err):
-            print("error: \(err)")
-        }
-    }
-}
