--- OmiseSwift/CardBrand.swift
@@ -1,11 +1,0 @@
-import Foundation
-
-public enum CardBrand: String {
-    case visa = "Visa"
-    case masterCard = "MasterCard"
-    case jcb = "JCB"
-    case amex = "American Express"
-    case diners = "Diners Club"
-    case discover = "Discover"
-    case maestro = "Maestro"
-}
