--- OmiseSwift/DisputeStatus.swift
@@ -1,9 +1,0 @@
-import Foundation
-
-public enum DisputeStatus: String {
-    case open
-    case pending
-    case won
-    case lost
-    case closed
-}
