--- OmiseSwift/RecipientType.swift
@@ -1,6 +1,0 @@
-import Foundation
-
-public enum RecipientType: String {
-    case individual
-    case corporation
-}
