--- OmiseSwift/Refund new.swift
@@ -1,73 +1,0 @@
-import Foundation
-
-
-public struct Refund: OmiseResourceObject {
-    public static let resourceInfo: ResourceInfo = ResourceInfo(path: "/refunds")
-    
-    public let object: String
-    public let location: String
-    
-    public let id: String
-    public let isLive: Bool
-    public var createdDate: Date
-    public let isDeleted: Bool
-    
-    public let value: Value
-    
-    public let charge: DetailProperty<Charge>
-    public let transaction: DetailProperty<Transaction>
-}
-
-extension Refund {
-    public init?(JSON json: Any) {
-        guard let json = json as? [String: Any], let omiseObjectProperties = Charge.parseOmiseResource(JSON: json) else {
-            return nil
-        }
-        
-        guard let value = Value(JSON: json), let charge = json["charge"].flatMap(DetailProperty<Charge>.init(JSON:)),
-            let transaction = json["transaction"].flatMap(DetailProperty<Transaction>.init(JSON:)) else {
-                return nil
-        }
-        (self.object, self.location, self.id, self.isLive, self.createdDate, self.isDeleted) = omiseObjectProperties
-        
-        self.value = value
-        self.charge = charge
-        self.transaction = transaction
-    }
-}
-
-
-public struct RefundParams: APIParams {
-    public var amount: Int64
-    public var void: Bool?
-    
-    public var json: JSONAttributes {
-        return Dictionary.makeFlattenDictionaryFrom([
-            "amount": amount,
-            "void": void,
-            ])
-    }
-}
-
-extension Refund: Creatable {
-    public typealias CreateParams = RefundParams
-}
-
-extension Refund: Listable { }
-extension Refund: Retrievable { }
-
-
-extension Charge {
-    public func listRefunds(using client: APIClient, params: ListParams? = nil, callback: @escaping Refund.ListRequest.Callback) -> Refund.ListRequest? {
-        return Refund.list(using: client, parent: self, params: params, callback: callback)
-    }
-    
-    public func retrieveRefund(using client: APIClient, id: String, callback: @escaping Refund.RetrieveRequest.Callback) -> Refund.RetrieveRequest? {
-        return Refund.retrieve(using: client, parent: self, id: id, callback: callback)
-    }
-    
-    public func createRefund(using client: APIClient, params: RefundParams, callback: @escaping Refund.CreateRequest.Callback) -> Refund.CreateRequest? {
-        return Refund.create(using: client, parent: self, params: params, callback: callback)
-    }
-}
-
