+++ OmiseSwift/Refund.swift
@@ -1,40 +1,51 @@
 import Foundation
 
-open class Refund: ResourceObject {
-    open override class var info: ResourceInfo {
-        return ResourceInfo(parentType: Charge.self, path: "/refunds")
-    }
+
+public struct Refund: OmiseResourceObject {
+    public static let resourceInfo: ResourceInfo = ResourceInfo(path: "/refunds")
     
-    public var amount: Int64? {
-        get { return get("amount", Int64Converter.self) }
-        set { set("amount", Int64Converter.self, toValue: newValue) }
-    }
+    public let object: String
+    public let location: String
     
-    public var currency: Currency? {
-        get { return get("currency", StringConverter.self).flatMap(Currency.init(code:)) }
-        set { set("currency", StringConverter.self, toValue: newValue?.code) }
-    }
+    public let id: String
+    public let isLive: Bool
+    public var createdDate: Date
+    public let isDeleted: Bool
     
-    public var charge: String? {
-        get { return get("charge", StringConverter.self) }
-        set { set("charge", StringConverter.self, toValue: newValue) }
-    }
+    public let value: Value
     
-    public var transaction: String? {
-        get { return get("transaction", StringConverter.self) ?? getChild("transaction", Transaction.self)?.id }
-        set { set("transaction", StringConverter.self, toValue: newValue) }
+    public let charge: DetailProperty<Charge>
+    public let transaction: DetailProperty<Transaction>
+}
+
+extension Refund {
+    public init?(JSON json: Any) {
+        guard let json = json as? [String: Any], let omiseObjectProperties = Charge.parseOmiseResource(JSON: json) else {
+            return nil
+        }
+        
+        guard let value = Value(JSON: json), let charge = json["charge"].flatMap(DetailProperty<Charge>.init(JSON:)),
+            let transaction = json["transaction"].flatMap(DetailProperty<Transaction>.init(JSON:)) else {
+                return nil
+        }
+        (self.object, self.location, self.id, self.isLive, self.createdDate, self.isDeleted) = omiseObjectProperties
+        
+        self.value = value
+        self.charge = charge
+        self.transaction = transaction
     }
 }
 
-public class RefundParams: Params {
-    public var amount: Int64? {
-        get { return get("amount", Int64Converter.self) }
-        set { set("amount", Int64Converter.self, toValue: newValue) }
-    }
+
+public struct RefundParams: APIParams {
+    public var amount: Int64
+    public var void: Bool?
     
-    public var void: Bool? {
-        get { return get("void", BoolConverter.self) }
-        set { set("void", BoolConverter.self, toValue: newValue) }
+    public var json: JSONAttributes {
+        return Dictionary.makeFlattenDictionaryFrom([
+            "amount": amount,
+            "void": void,
+            ])
     }
 }
 
@@ -45,34 +56,18 @@ extension Refund: Creatable {
 extension Refund: Listable { }
 extension Refund: Retrievable { }
 
+
 extension Charge {
-    public func listRefunds(using given: Client? = nil, params: ListParams? = nil, callback: @escaping Refund.ListOperation.Callback) -> Request<Refund.ListOperation.Result>? {
-        return Refund.list(using: given ?? attachedClient, parent: self, params: params, callback: callback)
+    public func listRefunds(using client: APIClient, params: ListParams? = nil, callback: @escaping Refund.ListRequest.Callback) -> Refund.ListRequest? {
+        return Refund.list(using: client, parent: self, params: params, callback: callback)
     }
     
-    public func retrieveRefund(using given: Client? = nil, id: String, callback: @escaping Refund.RetrieveOperation.Callback) -> Request<Refund.RetrieveOperation.Result>? {
-        return Refund.retrieve(using: given ?? attachedClient, parent: self, id: id, callback: callback)
+    public func retrieveRefund(using client: APIClient, id: String, callback: @escaping Refund.RetrieveRequest.Callback) -> Refund.RetrieveRequest? {
+        return Refund.retrieve(using: client, parent: self, id: id, callback: callback)
     }
     
-    public func createRefund(using given: Client? = nil, params: RefundParams, callback: @escaping Refund.CreateOperation.Callback) -> Request<Refund.RetrieveOperation.Result>? {
-        return Refund.create(using: given ?? attachedClient, parent: self, params: params, callback: callback)
+    public func createRefund(using client: APIClient, params: RefundParams, callback: @escaping Refund.CreateRequest.Callback) -> Refund.CreateRequest? {
+        return Refund.create(using: client, parent: self, params: params, callback: callback)
     }
 }
 
-func exampleRefund() {
-    let charge = Charge()
-    charge.id = "chrg_test_123"
-    
-    let params = RefundParams()
-    params.amount = 1_000_00 // 1,000.00 THB
-    params.void = true
-    
-    _ = Refund.create(parent: charge, params: params) { (result) in
-        switch result {
-        case let .success(refund):
-            print("created refund: \(refund.id)")
-        case let .fail(err):
-            print("error: \(err)")
-        }
-    }
-}
