--- OmiseSwift/TransactionType.swift
@@ -1,6 +1,0 @@
-import Foundation
-
-public enum TransactionType: String {
-    case debit
-    case credit
-}
