+++ OmiseSwift/Transfer.swift
@@ -1,135 +1,117 @@
 import Foundation
 
-public class Transfer: ResourceObject {
-    public override class var info: ResourceInfo { return ResourceInfo(path: "/transfers") }
+public struct Transfer: OmiseResourceObject {
+    public static let resourceInfo: ResourceInfo = ResourceInfo(path: "/transfers")
     
-    public var recipient: String? {
-        get { return get("recipient", StringConverter.self) ?? getChild("recipient", Recipient.self)?.id }
-        set { set("recipient", StringConverter.self, toValue: newValue) }
-    }
+    public let object: String
+    public let location: String
     
-    public var bankAccount: BankAccount? {
-        get { return getChild("bank_account", BankAccount.self) }
-        set { setChild("bank_account", BankAccount.self, toValue: newValue) }
-    }
+    public let id: String
+    public let isLive: Bool
+    public var createdDate: Date
+    public let isDeleted: Bool
     
-    public var sent: Bool? {
-        get { return get("sent", BoolConverter.self) }
-        set { set("sent", BoolConverter.self, toValue: newValue) }
-    }
+    public let bankAccount: BankAccount
+    
+    public let isSent: Bool
+    public let sentDate: Date?
+    public let isPaid: Bool
+    public let paidDate: Date?
+    
+    public let value: Value
+    public let fee: Value
 
-    public var paid: Bool? {
-        get { return get("paid", BoolConverter.self) }
-        set { set("paid", BoolConverter.self, toValue: newValue) }
-    }
-    
-    public var sentDate: Date? {
-        get { return get("sent_at", DateConverter.self) }
-        set { set("sent_at", DateConverter.self, toValue: newValue) }
-    }
-    
-    public var paidDate: Date? {
-        get { return get("paid_at", DateConverter.self) }
-        set { set("paid_at", DateConverter.self, toValue: newValue) }
-    }
-
-    public var amount: Int64? {
-        get { return get("amount", Int64Converter.self) }
-        set { set("amount", Int64Converter.self, toValue: newValue) }
-    }
-    
-    public var fee: Int64? {
-        get { return get("fee", Int64Converter.self) }
-        set { set("fee", Int64Converter.self, toValue: newValue) }
-    }
-    
-    public var currency: Currency? {
-        get { return get("currency", StringConverter.self).flatMap(Currency.init(code:)) }
-        set { set("currency", StringConverter.self, toValue: newValue?.code) }
-    }
-    
-    public var failureCode: String? {
-        get { return get("failure_code", StringConverter.self) }
-        set { set("failure_code", StringConverter.self, toValue: newValue) }
-    }
-    
-    public var failureMessage: String? {
-        get { return get("failure_message", StringConverter.self) }
-        set { set("failure_message", StringConverter.self, toValue: newValue) }
-    }
-    
-    public var transaction: String? {
-        get { return get("transaction", StringConverter.self) ?? getChild(("transaction"), Transaction.self)?.id }
-        set { set("transaction", StringConverter.self, toValue: newValue) }
-    }
+    public let transaction: DetailProperty<Transaction>?
 }
 
-public class TransferParams: Params {
-    public var amount: Int64? {
-        get { return get("amount", Int64Converter.self) }
-        set { set("amount", Int64Converter.self, toValue: newValue) }
-    }
-    
-    public var recipient: String? {
-        get { return get("recipient", StringConverter.self) }
-        set { set("recipient", StringConverter.self, toValue: newValue) }
+
+extension Transfer {
+    public init?(JSON json: Any) {
+        guard let json = json as? [String: Any],
+            let omiseObjectProperties = Charge.parseOmiseResource(JSON: json) else {
+                return nil
+        }
+
+        guard let bankAccount = json["bank_account"].flatMap(BankAccount.init(JSON:)),
+            let value = Value(JSON: json),
+            let fee = json["fee"] as? Int64, let currency = json["currency"].flatMap(CurrencyFieldConverter.convert(fromAttribute:)),
+            let isSent = json["sent"] as? Bool, let isPaid = json["paid"] as? Bool else {
+                return nil
+        }
+            
+        (self.object, self.location, self.id, self.isLive, self.createdDate, self.isDeleted) = omiseObjectProperties
+        self.bankAccount = bankAccount
+        self.isSent = isSent
+        self.isPaid = isPaid
+        self.value = value
+        self.fee = Value(currency: currency, amount: fee)
+        
+        self.transaction = json["transaction"].flatMap(DetailProperty<Transaction>.init(JSON:))
+        self.sentDate = json["sent_at"].flatMap(DateConverter.convert(fromAttribute:))
+        self.paidDate = json["paid_at"].flatMap(DateConverter.convert(fromAttribute:))
     }
 }
 
 
-public class TransferFilterParams: OmiseFilterParams {
-    public var created: DateComponents? {
-        get { return get("created", DateComponentsConverter.self) }
-        set { set("created", DateComponentsConverter.self, toValue: newValue) }
-    }
-    public var amount: Double? {
-        get { return get("amount", DoubleConverter.self) }
-        set { set("amount", DoubleConverter.self, toValue: newValue) }
+public struct TransferParams: APIParams {
+    public var amount: Int64
+    public var recipientID: String?
+    
+    public var json: JSONAttributes {
+        return Dictionary.makeFlattenDictionaryFrom([
+            "amount": amount,
+            "recipient": recipientID,
+            ])
     }
     
-    public var currency: Currency? {
-        get { return get("currency", CurrencyFieldConverter.self) }
-        set { set("currency", CurrencyFieldConverter.self, toValue: newValue) }
+    public init(amount: Int64, recipientID: String? = nil) {
+        self.amount = amount
+        self.recipientID = recipientID
+    }
+}
+
+public struct UpdateTransferParams: APIParams {
+    public var amount: Int64
+    
+    public var json: JSONAttributes {
+        return Dictionary.makeFlattenDictionaryFrom([
+            "amount": amount,
+            ])
     }
     
-    public var bankLastDigits: String? {
-        get { return get("bank_last_digits", StringConverter.self) }
-        set { set("bank_last_digits", StringConverter.self, toValue: newValue) }
+    public init(amount: Int64) {
+        self.amount = amount
     }
+}
+
+
+public struct TransferFilterParams: OmiseFilterParams {
+    public var created: DateComponents?
+    public var amount: Double?
+    public var currency: Currency?
+    public var bankLastDigits: LastDigits?
+    public var fee: Double?
+    public var isPaid: Bool?
+    public var paidDate: DateComponents?
+    public var isSent: Bool?
+    public var sentDate: DateComponents?
+    public var failureCode: String?
+    public var failureMessage: String?
     
-    public var fee: Double? {
-        get { return get("fee", DoubleConverter.self) }
-        set { set("fee", DoubleConverter.self, toValue: newValue) }
-    }
-    
-    public var paid: Bool? {
-        get { return get("paid", BoolConverter.self) }
-        set { set("paid", BoolConverter.self, toValue: newValue) }
-    }
-    
-    public var paidDate: DateComponents? {
-        get { return get("paid_at", DateComponentsConverter.self) }
-        set { set("paid_at", DateComponentsConverter.self, toValue: newValue) }
-    }
-    
-    public var sent: Bool? {
-        get { return get("sent", BoolConverter.self) }
-        set { set("sent", BoolConverter.self, toValue: newValue) }
-    }
-    
-    public var sentDate: DateComponents? {
-        get { return get("sent_at", DateComponentsConverter.self) }
-        set { set("sent_at", DateComponentsConverter.self, toValue: newValue) }
-    }
-    
-    public var failureCode: String? {
-        get { return get("failure_code", StringConverter.self) }
-        set { set("failure_code", StringConverter.self, toValue: newValue) }
-    }
-    
-    public var failureMessage: String? {
-        get { return get("failure_message", StringConverter.self) }
-        set { set("failure_message", StringConverter.self, toValue: newValue) }
+    public var json: JSONAttributes {
+        return Dictionary.makeFlattenDictionaryFrom([
+            "created": DateComponentsConverter.convert(fromValue: created),
+            "amount": amount,
+            "currency": currency?.code,
+            "bank_last_digits": bankLastDigits?.lastDigits,
+            "fee": fee,
+            "paid": isPaid,
+            "paid_at": DateComponentsConverter.convert(fromValue: paidDate),
+            "sent": isSent,
+            "sent_at": DateComponentsConverter.convert(fromValue: sentDate),
+            "failure_code": failureCode,
+            "failure_message": failureMessage,
+            ])
     }
 }
 
@@ -142,7 +124,7 @@ extension Transfer: Creatable {
 }
 
 extension Transfer: Updatable {
-    public typealias UpdateParams = TransferParams
+    public typealias UpdateParams = UpdateTransferParams
 }
 
 extension Transfer: Destroyable { }
